Fix file merge issue and reenable Merge...
This is caused when a builtin library is present in the merged file, but
not in the source. The merge succeeds because the merged file has the
same name as a builtin library. However, upon saving, the merged file
does not have the same library instance as the builtin and throws a
LoaderException